Moving things around in the configuration readers
