+++ include/configuration/xercesguard.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+namespace Signalbox {
+  namespace Configuration {
+    class XercesGuard {
+    public:
+      XercesGuard();
+      virtual ~XercesGuard();
+    };
+  }
+}
