--- include/xercesguard.hpp
@@ -1,9 +1,0 @@
-#pragma once
-
-namespace Signalbox {
-  class XercesGuard {
-  public:
-    XercesGuard();
-    virtual ~XercesGuard();
-  };
-}
