+++ src/CMakeLists.txt
@@ -1,12 +1,12 @@
 SET( srcs )
 
-LIST(APPEND srcs xercesguard.cpp)
+LIST(APPEND srcs configuration/xercesguard.cpp)
 LIST(APPEND srcs configuration/xercesstringdeleter.cpp)
 LIST(APPEND srcs configuration/utils.cpp)
+LIST(APPEND srcs configuration/configreader.cpp)
 LIST(APPEND srcs itemid.cpp)
 LIST(APPEND srcs digitaloutputpindata.cpp)
 LIST(APPEND srcs digitalinputpindata.cpp)
-LIST(APPEND srcs configreader.cpp)
 LIST(APPEND srcs signalheaddata.cpp)
 LIST(APPEND srcs trackcircuitmonitordata.cpp)
 LIST(APPEND srcs signalheadpins.cpp)
