--- src/xercesguard.cpp
@@ -1,15 +1,0 @@
-#include <iostream>
-
-#include "xercesc/util/PlatformUtils.hpp"
-
-#include "xercesguard.hpp"
-
-namespace Signalbox {
-  XercesGuard::XercesGuard() {
-    xercesc::XMLPlatformUtils::Initialize();
-  }
-
-  XercesGuard::~XercesGuard() {
-    xercesc::XMLPlatformUtils::Terminate();
-  }
-}
