+++ tst/xercesguardtest.cpp
@@ -1,12 +1,12 @@
 #include <boost/test/unit_test.hpp>
 
-#include "xercesguard.hpp"
+#include "configuration/xercesguard.hpp"
 
 BOOST_AUTO_TEST_SUITE( XercesGuard )
 
 BOOST_AUTO_TEST_CASE( ConstructAndDestruct )
 {
-  Signalbox::XercesGuard g;
+  Signalbox::Configuration::XercesGuard g;
   BOOST_CHECK( (&g) != nullptr );
   BOOST_TEST_CHECKPOINT("Constructed");
 }
